+++ heat_api_local_check.py
@@ -47,7 +47,7 @@ def check(args, tenant_id):
     if is_up:
         # only want to send other metrics if api is up
         metric('heat_api_local_response_time',
-               'uint32',
+               'double',
                '%.3f' % milliseconds,
                'ms')
